+++ workers/ohsome_quality_analyst/api/request_models.py
@@ -39,7 +39,7 @@ class BaseIndicator(BaseModel):
     include_svg: bool = False
     include_html: bool = False
     include_data: bool = False
-    flatten: bool = True
+    flatten: bool = False
 
     class Config:
         """Pydantic config class."""
@@ -56,7 +56,7 @@ class BaseReport(BaseModel):
     include_svg: bool = False
     include_html: bool = False
     include_data: bool = False
-    flatten: bool = True
+    flatten: bool = False
 
     class Config:
         """Pydantic config class."""
@@ -192,7 +192,7 @@ INDICATOR_EXAMPLES = {
             "fidField": "ogc_fid",
             "includeSvg": False,
             "includeHtml": False,
-            "flatten": True,
+            "flatten": False,
         },
     },
     "Custom AOI": {
@@ -286,7 +286,7 @@ INDICATOR_EXAMPLES = {
             },
             "includeSvg": False,
             "includeHtml": False,
-            "flatten": True,
+            "flatten": False,
         },
     },
 }
@@ -303,7 +303,7 @@ REPORT_EXAMPLES = {
             "fidField": "ogc_fid",
             "includeSvg": False,
             "includeHtml": False,
-            "flatten": True,
+            "flatten": False,
         },
     },
     "Custom AOI": {
@@ -327,7 +327,7 @@ REPORT_EXAMPLES = {
             },
             "includeSvg": False,
             "includeHtml": False,
-            "flatten": True,
+            "flatten": False,
         },
     },
 }